+++ CHANGELOG.md
@@ -40,3 +40,7 @@
 ## 1.2.0 Changed setTimeout to setInterval
 
 ## 1.2.1 Legal requested copy changes
+
+## 1.2.2 Bug Fixes
+ - Fix for Escape key not closing DNS popup (caused by onetrust-banner-sdk changes) - add our own keydown handler
+ - Fix some errors with the Display of on/off text in DNS popup for Targeting/Advertising Cookies in addition to slider